Improved nginx setup.

+++ scripts/setup-nginx.sh
@@ -1,7 +1,5 @@
 #!/bin/bash
 
-apt-install-if-needed nginx-full
-
 cat > /tmp/nginx.conf <<EOF
 user www-data;
 worker_processes 2;
@@ -76,6 +74,12 @@ server {
 }
 EOF
 
-sudo mv /tmp/nginx.conf /etc/nginx/nginx.conf
-sudo mv /tmp/taiga.conf /etc/nginx/sites-available/default
-sudo /etc/init.d/nginx restart
+apt-install-if-needed nginx-full
+
+if [ ! -e /etc/nginx/.taiga ]; then
+    touch /etc/nginx/.taiga
+
+    sudo mv /tmp/nginx.conf /etc/nginx/nginx.conf
+    sudo mv /tmp/taiga.conf /etc/nginx/sites-available/default
+    sudo /etc/init.d/nginx restart
+fi
